refactored keyboard compo and adjust css for responsive
