+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Hangman Game</title>
-    <link rel="stylesheet" href="app.css" />
+    <link rel="stylesheet" href="./css/app.css" />
     <link rel="icon" href="./img/hangman-favi.png" type="image/png" />
   </head>
   <body>
